added symmetry option to base model
